+++ src/communication/azext_communication/aaz/latest/communication/_create.py
@@ -19,6 +19,7 @@ class Create(AAZCommand):
 
     :example: Create a communication resource with tags
         az communication create -n ResourceName -g ResourceGroup --location global --data-location unitedstates --tags "{tag:tag}"
+
     """
 
     _aaz_info = {
@@ -106,6 +107,7 @@ class Create(AAZCommand):
         _args_schema = cls._args_schema
         _args_schema.data_location = AAZStrArg(
             options=["--data-location"],
+            required=True,
             arg_group="Properties",
             help="The location where the communication service stores its data at rest.",
         )