+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/storageappliance/_disable_remote_vendor_management.py
@@ -13,7 +13,7 @@ from azure.cli.core.aaz import *
 
 @register_command(
     "networkcloud storageappliance disable-remote-vendor-management",
-    is_experimental=True,
+    is_preview=True,
 )
 class DisableRemoteVendorManagement(AAZCommand):
     """Disable remote vendor management of the provided storage appliance.
@@ -23,9 +23,9 @@ class DisableRemoteVendorManagement(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2023-05-01-preview",
+        "version": "2023-07-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/storageappliances/{}/disableremotevendormanagement", "2023-05-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/storageappliances/{}/disableremotevendormanagement", "2023-07-01"],
         ]
     }
 
@@ -150,7 +150,7 @@ class DisableRemoteVendorManagement(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2023-05-01-preview",
+                    "api-version", "2023-07-01",
                     required=True,
                 ),
             }