+++ src/elastic-san/HISTORY.rst
@@ -3,6 +3,13 @@
 Release History
 ===============
 
+1.0.0b1
+++++++
+* Support private endpoint
+* Support 2022-12-01-preview api-version
+* BREAKING CHANGE: `--tags` has been removed from `az elastic-san volume-group/volume`
+* BREAKING CHANGE: `--size-gib` is required for `az elastic-san volume create`
+
 0.1.0
 ++++++
 * Initial release.