+++ src/interactive/HISTORY.rst
@@ -6,6 +6,7 @@ Release History
 upcoming
 +++++
 * Fix Scenario Idx out of bound when selecting recommended scenario
+* Add command skipped message in Scenario Execution Mode
 
 0.5.1
 +++++